requirements-backend.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pandas==1.3.5
numpy==1.21.4
ta==0.11.0
python-dateutil==2.8.2
pydantic==2.5.0
pyzmq==25.1.1
